server/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "uvicorn>=0.34.3",
    "sqlmodel>=0.0.14",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.2",
    "reportlab>=4.0.8",
    "python-dotenv>=1.0.0",
    "sqlalchemy-sqlitecloud>=0.1.2",
]
